requirements.txt
# ============================================
# DEPENDENCIAS DEL PROYECTO CANAL RELI
# Ubicación: C:\docker\projects\canal-reli\requirements.txt
# ============================================

# ============================================
# API DE ANTHROPIC (Claude)
# ============================================
anthropic==0.42.0

# ============================================
# PROCESAMIENTO DE DATOS
# ============================================
pandas>=2.0.0
openpyxl==3.1.2
pyyaml>=6.0

# ============================================
# AUDIO - WHISPER STT
# ============================================
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0

# ============================================
# GENERACIÓN DE PORTADAS (Pillow)
# Composición de imágenes, texto, gradientes
# ============================================
Pillow>=10.0.0

# ============================================
# CONFIGURACIÓN Y VALIDACIÓN
# ============================================
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# ============================================
# YOUTUBE API
# ============================================
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
python-dateutil>=2.8.2

# ============================================
# UTILIDADES
# ============================================
requests>=2.31.0

# ============================================
# DESARROLLO Y TESTING (opcional)
# ============================================
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0